+++ eslint.config.js
@@ -5,6 +5,6 @@ export default tseslint.config(
   eslint.configs.recommended,
   ...tseslint.configs.recommended,
   {
-    ignores: ['dist/**', 'example/**'],
+    ignores: ['dist/**', 'example/**', 'node_modules/**'],
   }
 )